draft refactor create link
